refactor: Move `Reflection`'s methods to their own modules
